% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{allele_flip}
\alias{allele_flip}
\title{Allele Flip Function for SNP Matching}
\usage{
allele_flip(sumstat, ref)
}
\arguments{
\item{sumstat}{A data frame containing the summary statistics data with at least
the columns "REF".}

\item{ref}{A data frame serving as the reference with at least the columns
"REF", "ALT", "BETA", and "SE".}
}
\value{
A modified version of \code{ref} where alleles have been flipped to match
\code{sumstat} and corresponding statistics (BETA and Z) have been updated accordingly.
}
\description{
This function modifies the reference dataset (\code{ref}) based on the summary
statistics dataset (\code{sumstat}) to ensure that the alleles match. If the alleles
do not match, the function will flip the alleles and corresponding statistics
in the reference dataset.
}
