% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{GWAS_QC}
\alias{GWAS_QC}
\title{GWAS Quality Control Function}
\usage{
GWAS_QC(
  dat,
  MAF_threshold,
  rm_strand = TRUE,
  rm_multi = TRUE,
  rm_indel = TRUE,
  rm_mhc = TRUE,
  rm_maf = TRUE
)
}
\arguments{
\item{dat}{A data frame containing the GWAS data. Columns "CHR", "POS", "CHR_POS",
"REF", "ALT", and (optionally) "MAF" are required.}

\item{MAF_threshold}{The threshold for Minor Allele Frequency filtering. SNPs
with MAF below this threshold or above (1 - threshold) will be removed.}
}
\value{
A filtered data frame after performing the quality control steps.
}
\description{
This function performs several quality control steps on GWAS data:
\enumerate{
\item Removes strand ambiguous variants.
\item Removes multi-allelic variants.
\item Excludes the MHC complex region.
\item Filters SNPs based on Minor Allele Frequency (MAF).
}
}
